Trivial warning fixes
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@17596 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/mibII/system_mib.c
@@ -103,7 +103,7 @@ system_parse_config_string(const char *token, char *cptr,
 {
     if (strlen(cptr) >= size) {
 	netsnmp_config_error("%s token too long (must be < %lu):\n\t%s",
-			     token, size, cptr);
+			     token, (unsigned long)size, cptr);
     }
 
     if (*token == 'p' && strcasecmp(token + 1, name) == 0) {
